change int to nint in grid_mod
git-svn-id: svn://hydrogene.lct.jussieu.fr/champ@327 36b7402c-5635-0410-8f0e-a47a0fc0da44

+++ qmc/grid_mod.f90
@@ -134,7 +134,7 @@ module grid_mod
   call require (lhere, 'grid_r_max > 0', grid_r_max > 0.d0 )
   call require (lhere, 'grid_r_min >= 0', grid_r_min >= 0.d0 )
 
-  grid_r_nb = int((grid_r_max-grid_r_min)/grid_r_step) + 1
+  grid_r_nb = nint((grid_r_max-grid_r_min)/grid_r_step) + 1
   call require (lhere, 'grid_r_nb > 0', grid_r_nb > 0 )
 
   write (6,'(a)') ' Parameters for radial grid:'
